fix bug 2826

+++ scilab/modules/gui/sci_gateway/c/sci_helpbrowser.c
@@ -76,11 +76,11 @@ int sci_helpbrowser(char *fname,unsigned long fname_len)
 
       if (helpAdr == NULL) /* No toolboxes loaded */
         {
-          launchHelpBrowser(NULL, nbRowHelp*nbColHelp, getStringMatrixFromStack((int)languageAdr)[0]);
+          launchHelpBrowser(NULL, nbRowHelp*nbColHelp, getStringMatrixFromStack(languageAdr)[0]);
         }
       else
         {
-          launchHelpBrowser(getStringMatrixFromStack((int)helpAdr), nbRowHelp*nbColHelp, getStringMatrixFromStack((int)languageAdr)[0]);
+          launchHelpBrowser(getStringMatrixFromStack((int)helpAdr), nbRowHelp*nbColHelp, getStringMatrixFromStack(languageAdr)[0]);
         }
     }
   else if (Rhs == 4)
@@ -132,11 +132,11 @@ int sci_helpbrowser(char *fname,unsigned long fname_len)
       
       if (helpAdr == NULL) /* No toolboxes loaded */
         {
-          searchKeyword(NULL, nbRowHelp*nbColHelp, getStringMatrixFromStack((int)keywordAdr)[0], getStringMatrixFromStack((int)languageAdr)[0], *istk(fullTextAdr)==1);
+          searchKeyword(NULL, nbRowHelp*nbColHelp, getStringMatrixFromStack((int)keywordAdr)[0], getStringMatrixFromStack(languageAdr)[0], *istk(fullTextAdr)==1);
         }
       else
         {
-          searchKeyword(getStringMatrixFromStack((int)helpAdr), nbRowHelp*nbColHelp, getStringMatrixFromStack((int)keywordAdr)[0], getStringMatrixFromStack((int)languageAdr)[0], *istk(fullTextAdr)==1);
+          searchKeyword(getStringMatrixFromStack((int)helpAdr), nbRowHelp*nbColHelp, getStringMatrixFromStack((int)keywordAdr)[0], getStringMatrixFromStack(languageAdr)[0], *istk(fullTextAdr)==1);
         }
     }
   else
